chore: fix clippy 1.95 lints
Address 16 clippy 1.95 warnings/errors across the H.265 crate by
rewriting the flagged code idiomatically. No crate-wide allow shotgun;
no per-call-site allow either — every issue had a clean rewrite.

Highlights:
* `doc_lazy_continuation` on the encoder rustdoc list item: indented
  the multi-line continuation under the `+` sub-bullet so markdown
  treats it as a single paragraph (lib.rs).
* `manual_clamp` → `clamp(1, 255)` in scaling_list scaling-list-data
  delta walker (§7.4.5).
* `needless_range_loop` rewritten with iterator+enumerate for the
  deblock BETA_TABLE/TC_TABLE/CHROMA_QP_I2C tests and for the
  inverse-diag-scan invariance test in scaling_list.
* `if_same_then_else` in reference_clip Main10 fixture builder:
  collapsed the two `Some(out)` arms into a single `||` condition
  (the first branch only existed for its ffmpeg side-effect).
* `never_loop` in encoder_roundtrip: loop body always broke or
  panicked on the first iteration → replaced with a plain `match`.
* `clippy --fix` auto-fixes folded in across cabac, ctu,
  encoder/{bit_writer,cabac_writer,residual_writer}, intra_pred,
  sps, transform: `let_and_return`, `unnecessary_cast`,
  `collapsible_if`. Each is a no-op semantic change.

cargo clippy --all-targets -- -D warnings: clean.
cargo fmt --check: clean.
cargo test: 125 passed, 0 failed, 3 ignored.

+++ src/encoder/cabac_writer.rs
@@ -102,7 +102,7 @@ impl<'a> CabacWriter<'a> {
     pub fn encode_flush(&mut self) {
         self.ivl_curr_range = 2;
         self.renormalise();
-        self.put_bit(((self.ivl_low >> 9) & 1) as u32);
+        self.put_bit((self.ivl_low >> 9) & 1);
         // Write 2 bits: top bit of the remaining 8 is in bit 8, then bit 7;
         // spec combines as `((ivl_low >> 7) & 3) | 1`.
         let bits = ((self.ivl_low >> 7) & 3) | 1;
